+++ tests/testthat/test-select-helpers.R
@@ -323,7 +323,7 @@ test_that("all_of() and any_of() check their inputs", {
   expect_error(vars_select(letters, any_of(na_chr)), "missing")
 })
 
-test_that("starts_with() and ends_with() accept length > 1 vectors (#50)", {
+test_that("matchers accept length > 1 vectors (#50)", {
   expect_identical(
     vars_select(names(iris), starts_with(c("Sep", "Petal"))),
     vars_select(names(iris), starts_with("Sep") | starts_with("Petal"))
@@ -331,5 +331,13 @@ test_that("starts_with() and ends_with() accept length > 1 vectors (#50)", {
   expect_identical(
     vars_select(names(iris), ends_with(c("gth", "Width"))),
     vars_select(names(iris), ends_with("gth") | ends_with("Width"))
+  )
+  expect_identical(
+    vars_select(names(iris), contains(c("epal", "eta"))),
+    vars_select(names(iris), contains("epal") | contains("eta")),
+  )
+  expect_identical(
+    vars_select(names(iris), matches(c("epal", "eta"))),
+    vars_select(names(iris), matches("epal") | contains("eta")),
   )
 })